backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
pdfminer.six==20231228
python-docx==1.1.2
rapidfuzz==3.10.0